src/server/vector_db.py
# Previous contents of contextual_vector_db.py moved here
import os
import pickle
import json
import numpy as np
import voyageai
from typing import List, Dict, Any, Optional
from tqdm import tqdm
from openai import OpenAI
import threading
import time
from concurrent.futures import ThreadPoolExecutor, as_completed
from dotenv import load_dotenv

# Rest of the contextual_vector_db.py content...
